requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.7.1
dnspython==2.4.2
requests==2.32.3